Refactor sanity classes for use in all tests.
